Moved spirals spot picking to a new file
