+++ rsoxs_scans/spreadsheetManipulation.py
@@ -1,0 +1,13 @@
+
+
+## New spreadsheet 
+
+def loadSpreadsheet_Local():
+    ## TODO: copy over eliot's code piece by piece and make comments
+    ## TODO: Bar sheet should remain same but rename to samples maybe (see my partway done spreadsheet in rsoxs package)
+    ## TODO: redo acquisitions
+
+
+
+def sanitizeAcquisitions():
+    ##
